added new music and shuffle system.
